+++ apps/app-sample/routes/index.js
@@ -12,7 +12,6 @@ module.exports = ({ app, routePrefix }) => {
     router.use('/', require('./base')), // http://127.0.0.1:3000/api/app-sample/
     router.use('/categories', require('./categories')), // http://127.0.0.1:3000/api/app-sample/categories/
     router.use('/webhooks', require('./webhooks')),
-    router.use('/mongo-demo', require('./mongo-demo')), // for demos
     router.use('/sse', require('./sse')),
     router.use('/tests', require('./tests')), // for tests
     router.use('/webpush', require('./webpush')),