+++ apps/app-sample/routes/tests.js
@@ -130,7 +130,7 @@ module.exports = express.Router()
     res.send("ws broadcast")
   })
 
-  // TODO /esm/upload-fe-test.js
+  // TODO /esm/upload-fe-testing.js
   // test uploads
   // body action: 'read' | 'write', filename: 'my-file.txt', bucket: 'bucket name'
   .post('/gcp-sign', gcp.getSignedUrl)